validate all exported paths in satndalone and maya working
