+++ projects/framework-common-extensions/tool-configs/standalone-file-publisher.yaml
@@ -42,6 +42,13 @@ engine:
         options:
           export_destination: "~/Desktop/myPublishedFile.png"
 
+  # Common validator check all exported paths exists.
+  - type: plugin
+    tags:
+      - validator
+    plugin: exported_paths_validator
+    ui: validator_label
+
 
   # Publish to ftrack.
   - publish_to_ftrack
